separated filter related code to custom useTaskFilter() hook
