+++ src/features/tasks/components/kanban-task/KanbanTask.tsx
@@ -1,11 +1,9 @@
 import { MutableRefObject } from 'react';
 import PrimaryButton from '../../../../components/button/primary-button/PrimaryButton';
 import Dialog from '../../../../components/dialog/Dialog';
-import TextField from '../../../../components/input/TextField';
-import SelectField from '../../../../components/select/SelectField';
-import { priorityLevels } from '../../constants';
 import { Task, TeamMember } from '../../types';
-import useHandleTask from '../../hooks/useHandleTask';
+import useEditTask from '../../hooks/useEditTask';
+import EditableTaskContent from '../editable-task-content/EditableTaskContent';
 
 type KanbanTaskProps = {
     tasks: Task[];
@@ -35,7 +33,7 @@ export default function KanbanTask({
         editedTask,
         startTaskEdit,
         closeTaskEdit,
-    } = useHandleTask({
+    } = useEditTask({
         tasks,
         task,
         handleSetTasks,
@@ -58,74 +56,16 @@ export default function KanbanTask({
                 closeDialog={closeDialog}
                 className='kanban-task__dialog'
             >
+                <h1>Edit task</h1>
+
                 {isEditing ? (
-                    <div className='kanban-task-edit'>
-                        <div className='kanban-task-edit__item'>
-                            <TextField
-                                label='Name'
-                                inputId='task-name-edit'
-                                value={editedTask.name}
-                                onChange={(newValue) => handleTaskPropertyChange('name', newValue)}
-                            />
-                        </div>
-
-                        <div className='kanban-task-edit__item'>
-                            <label htmlFor='due-date-edit'>Due date</label>
-
-                            <input
-                                type='date'
-                                id='due-date-edit'
-                                value={editedTask.dueDate.toISOString().split('T')[0]}
-                                onChange={(e) => {
-                                    handleSetSelectedDueDate(new Date(e.target.value));
-
-                                    if (e.target.value === '') {
-                                        handleSetSelectedDueDate(new Date());
-                                    }
-
-                                    handleTaskPropertyChange('dueDate', new Date(e.target.value));
-                                }}
-                            />
-                        </div>
-
-                        <div className='kanban-task-edit__item'>
-                            <SelectField
-                                label='Priority Level'
-                                value={editedTask.priorityLevel}
-                                selectId='priority-edit-select'
-                                onChange={(newValue) => handleTaskPropertyChange('priorityLevel', newValue as string)}
-                                options={priorityLevels.map((priorityLevel) => ({
-                                    value: priorityLevel,
-                                    label: priorityLevel,
-                                }))}
-                            />
-                        </div>
-
-                        <div className='kanban-task-edit__item'>
-                            <SelectField
-                                label='Assigned Team Member'
-                                value={editedTask.assignedTeamMember ? editedTask.assignedTeamMember.id : ''}
-                                selectId='assigned-team-member-select-edit'
-                                options={[
-                                    { value: '', label: 'Select Team Member' },
-                                    ...users.map((teamMember) => ({
-                                        value: teamMember.id,
-                                        label: teamMember.name,
-                                    })),
-                                ]}
-                                onChange={(newValue) => {
-                                    const selectedTeamMember =
-                                        users.find((member) => member.id === Number(newValue)) ?? null;
-                                    handleTaskPropertyChange('assignedTeamMember', selectedTeamMember);
-                                }}
-                            />
-                        </div>
-
-                        <PrimaryButton
-                            onClick={closeTaskEdit}
-                            content={<>FInish editing</>}
-                        />
-                    </div>
+                    <EditableTaskContent
+                        users={users}
+                        editedTask={editedTask}
+                        handleTaskPropertyChange={handleTaskPropertyChange}
+                        handleSetSelectedDueDate={handleSetSelectedDueDate}
+                        closeTaskEdit={closeTaskEdit}
+                    />
                 ) : (
                     <>
                         <h1>{task.name}</h1>
@@ -135,15 +75,8 @@ export default function KanbanTask({
                             <div>Assigned Team Member: {task.assignedTeamMember.name}</div>
                         ) : null}
 
-                        <PrimaryButton
-                            onClick={startTaskEdit}
-                            content={<>Edit</>}
-                        />
-
-                        <PrimaryButton
-                            onClick={handleDeleteTask}
-                            content={<>Delete</>}
-                        />
+                        <PrimaryButton onClick={startTaskEdit}>Edit</PrimaryButton>
+                        <PrimaryButton onClick={handleDeleteTask}>Delete</PrimaryButton>
                     </>
                 )}
             </Dialog>
