+++ src/features/tasks/components/kanban-task/kanban-task.scss
@@ -1,5 +1,5 @@
 .kanban-task {
-    border: 2px solid black;
+    border: 1px solid black;
     margin: 2rem;
     padding: 1rem;
     cursor: pointer;
